refactor: strategy code splitting refactor: add "main()" function to src/index.ts
